Implement EQ alerts worker
