+++ cache.json
@@ -1,1 +1,1 @@
-{"time":"2018-07-06T21:05:00.000Z","when":"2018-07-06T22:00:00.000Z","eqs":[{"name":"異界に紡がれし知の化身","jpName":"異界に紡がれし知の化身","ship":1},{"name":"異界に紡がれし知の化身","jpName":"異界に紡がれし知の化身","ship":2},{"name":"異界に紡がれし知の化身","jpName":"異界に紡がれし知の化身","ship":3},{"name":"異界に紡がれし知の化身","jpName":"異界に紡がれし知の化身","ship":4},{"name":"異界に紡がれし知の化身","jpName":"異界に紡がれし知の化身","ship":5},{"name":"異界に紡がれし知の化身","jpName":"異界に紡がれし知の化身","ship":6},{"name":"異界に紡がれし知の化身","jpName":"異界に紡がれし知の化身","ship":7},{"name":"異界に紡がれし知の化身","jpName":"異界に紡がれし知の化身","ship":8},{"name":"異界に紡がれし知の化身","jpName":"異界に紡がれし知の化身","ship":9},{"name":"異界に紡がれし知の化身","jpName":"異界に紡がれし知の化身","ship":10}]}
+2018-07-11T20:05:04.000Z
